fix: o worker nao recebia a composicao da API — 336 BankError/24h em producao
